% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_helper.R
\name{get_src_table}
\alias{get_src_table}
\title{Get Source Table}
\usage{
get_src_table(
  path,
  FILE_pattern = ".xlsx|.xls|.csv",
  ID_pattern = "ID",
  DATE_pattern = "DATE|VISITNO",
  IS_overlap_list = NULL,
  WINDOW_list = NULL,
  ID_usr_list = NULL,
  DATE_usr_list = NULL,
  file = NULL
)
}
\arguments{
\item{path}{The path to the directory containing the data files.}

\item{FILE_pattern}{A regular expression pattern specifying the types of files to be read in. Default is ".xlsx|.xls|.csv".}

\item{ID_pattern}{A regular expression pattern specifying the potential ID variables in the data. Default is "ID".}

\item{DATE_pattern}{A regular expression pattern specifying the potential DATE variables in the data. Default is "DATE|VISITNO".}

\item{IS_overlap_list}{A list of Boolean values specifying whether or not the DATE variables overlap for each file. Default is NULL.}

\item{WINDOW_list}{A list of numric time windows for matching the DATE variables for each file. Default is NULL.}

\item{ID_usr_list}{A list of user-specified ID variable names for each file.}

\item{DATE_usr_list}{A list of user-specified DATE variable names for each file.}

\item{file}{A path to a file where the source table will be saved as a CSV file. If NULL, it will not be saved locally.}
}
\value{
A data frame summarizing the variables available in each file, with information on potential ID and DATE variables and the ID and DATE variables to be used for merging.
}
\description{
This function reads in data files from a specified directory and generates a source table
that summarizes the variables available in each file.
The source table includes information on the file name, variables in the file,
potential ID variables, potential DATE variables, and the ID and DATE variables to be used for merging.
The function also allows users to input merge options,
such as whether or not the DATE variables overlap and the time window for matching.
Users can also specify the names of the ID and DATE variables they wish to use for merging.
}
\examples{
\dontrun{
get_src_table("path/to/data/directory")
}
}
