% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_helper.R
\name{get_window_bound}
\alias{get_window_bound}
\title{Get the bounds for the window timeline.}
\usage{
get_window_bound(original_date, bound_date, is_left, ovlp, window_len)
}
\arguments{
\item{original_date}{The original date for which to calculate the window bound.}

\item{bound_date}{The bound date for the window, which can be specified as a specific date or a number of days.}

\item{is_left}{A logical value indicating whether the left or right bound should be returned.}

\item{ovlp}{A logical value indicating whether the data is overlapping or not. If the data is overlapping, the window bounds will be calculated using only the window length. If the data is not overlapping, the minimum of the bound date and today's date minus half the window length will be used.}

\item{window_len}{The length of the window in days.}
}
\value{
The left or right window bound for the given date.
}
\description{
This function calculates the window bounds for a given date.
The window bounds are used to define the time range for longitudinal data.
The function takes the original date, the bound date, the overlap flag,
and the window length as inputs, and returns the left or right window bound.
}
\examples{
\dontrun{
get_window_bound("2021-01-01", "2022-01-01", is_left = TRUE, ovlp = TRUE, window_len = 365)
}

}
