% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_helper.R
\name{ordered_match}
\alias{ordered_match}
\title{Match the first occurring item.}
\usage{
ordered_match(usr_list, file_list, default_list)
}
\arguments{
\item{usr_list}{A list of variables that the user is interested in finding a match for.}

\item{file_list}{A list of variables that are available in a file.}

\item{default_list}{A character vector of default variables to use if no matching variables are found.}
}
\value{
The matched variable.
}
\description{
This function finds the first variable in usr_list that matches one of the variables in file_list.
If none of the variables in the usr_list match any of the variables in the file_list,
the function returns the first variable in the default_list
}
\examples{
ordered_match(c("A", "B"), "C; D; A", c("X", "Y", "Z"))
}
