% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adm_summary.R
\name{plot.ADMerge_res}
\alias{plot.ADMerge_res}
\title{Plot function for ADMerge analysis results}
\usage{
\method{plot}{ADMerge_res}(res, distn, group, baseline = TRUE, ...)
}
\arguments{
\item{res}{The result object from the ADMerge function.}

\item{distn}{The name of the variable to plot the distribution of.}

\item{group}{The name of the variable to group the distribution by.}

\item{baseline}{A boolean indicating whether to include only the baseline data in the plot.}

\item{...}{Additional parameters for the function.}
}
\value{
A bar plot of the distribution of the given variable among different groups in the merged dataset.
}
\description{
This function generates a bar plot of the distribution of a given variable among different groups in the merged dataset generated by the ADMerge function.
}
\examples{
\dinttun{
plot.ADMerge_res(res, "AGE", "SEX")
}

}
